% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AmatrixHD}
\alias{AmatrixHD}
\title{The matrix A in the equation Ax=b for 1 haploid and 1 diploid chromosome}
\usage{
AmatrixHD(pop_allele_freqs, genotypes)
}
\arguments{
\item{pop_allele_freqs}{A numeric vector of population allele frequencies for each SNP}

\item{genotypes}{An integer matrix of genotype calls for a pair of isolates. Each coloumn represents and isolate and each row represents a SNP.}
}
\description{
The matrix A in the equation Ax=b for 1 haploid and 1 diploid chromosome
}
