% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_genes.R
\docType{data}
\name{annotation_genes}
\alias{annotation_genes}
\title{Plasmodium Falciparum Gene Annotation Dataset}
\format{A data frame with 6 columns of information
 \describe{
\item{chr}{Chromosomes}
\item{start}{Base-pair positions of start of genes}
\item{end}{Base-pair positions of end of genes}
\item{strand}{The positive or negaitve gene strand}
\item{name}{Gene name, commonly abbreviated}
\item{gene_id}{Gene id}
}}
\usage{
annotation_genes
}
\description{
Gene annotations for the reference genome 3D7 were downloaded from
\url{http://www.plasmodb.org/common/downloads/Current_Release/Pfalciparum3D7/gff/data/},
release PlasmoDB-28, last modified 23/03/2016.
}
\keyword{datasets}

