% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_are_colours.R
\name{areColors}
\alias{areColors}
\title{Internal function}
\usage{
areColors(x)
}
\arguments{
\item{x}{vector of length 1 or higher containing numeric or character values for colours}
}
\description{
\code{areColors()} checks if colour names are valid
}
