% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BAFemprobs}
\alias{BAFemprobs}
\title{The emission probability for isolate i, SNP j in state k.}
\usage{
BAFemprobs(BAFvalue, meanBAF0, sdBAF0, state0, pfb)
}
\arguments{
\item{BAFvalue}{A vector of length 1 with the BAF value for isolate i, SNP j}

\item{meanBAF0}{A vector of possible major clone proportions}

\item{sdBAF0}{A vector of standard deviations for; one for each major clone proportion}

\item{state0}{The IBD state}

\item{pfb}{The population BAF for SNP j}
}
\description{
The emission probability for isolate i, SNP j in state k.
}

