% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bVectorDD}
\alias{bVectorDD}
\title{The vector b in the equation Ax=b for 2 diploid chromosomes}
\usage{
bVectorDD(genotypes)
}
\arguments{
\item{genotypes}{An integer matrix of genotype calls for a pair of isolates. Each coloumn represents and isolate and each row represents a SNP.}
}
\description{
The vector b in the equation Ax=b for 2 diploid chromosomes
}
