% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bVectorHD}
\alias{bVectorHD}
\title{The vector b in the equation Ax=b for 1 haploid chromosome and 1 diploid chromosome}
\usage{
bVectorHD(genotypes)
}
\arguments{
\item{genotypes}{An integer matrix of genotype calls for a pair of isolates. Each coloumn represents and isolate and each row represents a SNP.}
}
\description{
The vector b in the equation Ax=b for 1 haploid chromosome and 1 diploid chromosome
}
