% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_em_probs}
\alias{calculate_em_probs}
\title{The emission probabilities for isolate i calculated over all SNPs and states}
\usage{
calculate_em_probs(BAFmatrix, alleleFreq, BAFmean, BAFsd, positionBP, bin)
}
\arguments{
\item{BAFmatrix}{A vector of BAF values for isolate i}

\item{alleleFreq}{A vector of population BAF values}

\item{BAFmean}{A vector of possible major clone proportions}

\item{BAFsd}{A vector of standard deviations for; one for each major clone proportion}

\item{positionBP}{A vector of genetic map positions for all SNPs in base-pairs}

\item{bin}{A matrix of bins with parameter estimates (proportion estimates with corresponding SD)}
}
\description{
The emission probabilities for isolate i calculated over all SNPs and states
}

