% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_gamma_w}
\alias{calculate_gamma_w}
\title{Calculate gamma}
\usage{
calculate_gamma_w(alpha, beta)
}
\arguments{
\item{alpha}{A matrix of alpha values calculated from the forward algorithm}

\item{beta}{A matrix of beta values calculated from the backward algorithm}
}
\description{
Calculate gamma
}

