% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_logLikelihood_w}
\alias{calculate_logLikelihood_w}
\title{Calculate the log-likelihood}
\usage{
calculate_logLikelihood_w(scale)
}
\arguments{
\item{scale}{A vector of scaling values used to scale alpha}
}
\description{
Calculate the log-likelihood
}

