% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_viterbi_w}
\alias{calculate_viterbi_w}
\title{Calculate the Viterbi sequence}
\usage{
calculate_viterbi_w(emissionProbs, initalProbs, positionBP, transParam)
}
\arguments{
\item{emissionProbs}{A numeric matrix of pre-calculated emission probabilities for each state and SNP}

\item{initalProbs}{A numeric vector of initial state probabilities. One probability per state}

\item{positionBP}{A vector of genetic map positions in base-pairs}

\item{transParam}{A matrix of transition probabilities}
}
\description{
Calculate the Viterbi sequence
}

