% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_xi}
\alias{calculate_xi}
\title{Calculate the Xi}
\usage{
calculate_xi(emissionProbs, positionBP, transParam, alpha, beta)
}
\arguments{
\item{emissionProbs}{A numeric matrix of pre-calculated emission probabilities for each state and SNP}

\item{positionBP}{A vector of genetic map positions in base-pairs}

\item{transParam}{A matrix of transition probabilities}

\item{alpha}{A matrix of alpha values calculated from the forward algorithm}

\item{beta}{A matrix of alpha values calculated from the backward algorithm}
}
\description{
Calculate the Xi
}

