% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateGamma}
\alias{calculateGamma}
\title{Calculate gamma}
\usage{
calculateGamma(number_states, initial_prob, meiosis, number_snps, genotypes,
  pop_allele_freqs, positions_cM, error, gender_1, gender_2)
}
\arguments{
\item{number_states}{Integer. The number of IBD states in the model}

\item{initial_prob}{A numeric vector containing the initial state probabilities}

\item{meiosis}{Integer. The number of meiosis separating the two isolates}

\item{number_snps}{Integer. The number of SNPs}

\item{genotypes}{A integer martix containing the genotype calls for a pair of isolates}

\item{pop_allele_freqs}{A numeric vector of population allele frequencies}

\item{positions_cM}{A numeric vector of SNP genetic map positions in cM}

\item{error}{Numeric. The genotype error rate}

\item{gender_1}{Integer. The MOI estimate of isolate 1}

\item{gender_2}{Integer. The MOI estimate of isolate 2}
}
\description{
Calculate gamma
}

