% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_meiosis.R
\name{calculateMeiosis}
\alias{calculateMeiosis}
\title{Estimation of Meiosis}
\usage{
calculateMeiosis(omega.0, omega.1, omega.2)
}
\arguments{
\item{omega.0}{A numeric value between 0 and 1 representing the pop_allele_freqs of sharing 0 alleles IBD. The sum of omega.0, omega.1 and omega.2 should equal 1.}

\item{omega.1}{A numeric value between 0 and 1 representing the pop_allele_freqs of sharing 1 alleles IBD.}

\item{omega.2}{A numeric value between 0 and 1 representing the pop_allele_freqs of sharing 2 alleles IBD.}
}
\value{
The number of meiosis separating the pair of isoaltes.
}
\description{
calculateMeiosis() estimates the number of meiosis separating a pair of isolates given the global IBD pop_allele_freqs estimates.
This method is described in Purcell et al (2007).
}

