% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateMissingness}
\alias{calculateMissingness}
\title{Calculate Missingness Proportions}
\usage{
calculateMissingness(genotypes)
}
\arguments{
\item{genotypes}{An integer matrix of genotype data of the form -1, 0, 1 and 2 representing missing genotypes, homozygous reference, heterozygous and homozygous alternative respectively.}
}
\value{
A vector of length \code{n} where \code{n} is the number of columns in \code{genotypes}.
}
\description{
Calculates the proportion of missing data for each SNPs or each isolate where missing values are denoted by -1. 
Missing values are calculated for each column of \code{genotypes} (where columns are isolates and rows are SNPs),
however \code{genotypes} can be transposed to calculate missingness proportions for SNPs.
}

