% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculatePopAlleleFreq}
\alias{calculatePopAlleleFreq}
\title{Calculate Allele Frequencies for SNPs from Genotype Data}
\usage{
calculatePopAlleleFreq(genotypes, moi)
}
\arguments{
\item{genotypes}{An integer matrix of genotype data of the form -1, 0, 1 and 2 representing missing genotypes, homozygous reference, 
heterozygous and homozygous alternative respectively. Each column of \code{genotypes} represents a unique isolate and each row of \code{genotypes}
represents a unique SNP.}

\item{moi}{An integer vector of multiplicity of infection (MOI) estimates for each isoalte. Isolate MOI estimates should be ordered such that value \code{n} of 
\code{moi} corresponds to column \code{n} of \code{genotypes}.}
}
\description{
\code{calculatePopAlleleFreq} calculates reference allele frequencies for each SNP given genotype data.
}

