% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ibd_i_clusters.R
\name{clusterSummary}
\alias{clusterSummary}
\title{A function that prints summary information of clusters}
\usage{
clusterSummary(cluster.list)
}
\arguments{
\item{cluster.list}{a list of isolate groups where unique elements correspond to isolates in unique clusters}
}
\description{
A function that prints summary information of clusters
}
