% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{emissionProbDD}
\alias{emissionProbDD}
\title{The emission probabilities for 2 diploid chromosomes}
\usage{
emissionProbDD(pop_allele_freq, genotype_1, genotype_2, ibd)
}
\arguments{
\item{pop_allele_freq}{The population allele frequency for SNP i. This corresponds to the reference allele}

\item{genotype_1}{The genotype for isolate 1 from the pair for SNP i}

\item{genotype_2}{The genotype for isolate 2 from the pair for SNP i}

\item{ibd}{The IBD state}
}
\description{
The emission probabilities for 2 diploid chromosomes
}
