% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{emissionProbMissingGeno}
\alias{emissionProbMissingGeno}
\title{Calculating the emission probability sumation when missing genotype calls present}
\usage{
emissionProbMissingGeno(pop_allele_freq, genotype_1, genotype_2, error,
  gender_1, gender_2, ibd_j)
}
\arguments{
\item{pop_allele_freq}{The population allele frequency of SNP t}

\item{genotype_1}{The genotype of isolate 1 at SNP t}

\item{genotype_2}{The genotype of isolate 2 at SNP t}

\item{error}{The genotype error rate}

\item{gender_1}{The MOI estimate of isolate 1}

\item{gender_2}{The MOI estimate of isolate 2}

\item{ibd_j}{The IBD state}
}
\description{
Calculating the emission probability sumation when missing genotype calls present
}
