% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{genotypeErrorD}
\alias{genotypeErrorD}
\title{The genotyping error probability for 1 diploid chromosome}
\usage{
genotypeErrorD(truth, observed, error)
}
\arguments{
\item{truth}{The true genotype}

\item{observed}{The observed genotype}

\item{error}{The genotype error rate}
}
\description{
The genotyping error probability for 1 diploid chromosome
}

