% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_get_clusters.R
\name{getClusters}
\alias{getClusters}
\title{Between Isolates Clusters}
\usage{
getClusters(ped.genotypes, locus.matrix, snp = NULL, interval = NULL,
  prop = 0, hclust = FALSE)
}
\arguments{
\item{ped.genotypes}{a list containing 2 objects. See the \code{Value} description in \code{\link{between_data_processing}} for more details on this input.
Note the family IDs and isolate IDs in obeject 1 of this list must match the family IDs and isolate IDs in the header of object 2 of this list.}

\item{locus.matrix}{a data frame containing the binary IBD information for each SNP and each pair. 
See the returned \code{value} in \code{\link{between_locus_matrix}} for more details.}

\item{snp}{a character string of length 1 giving the name of a SNP to find clusters of isolates who are IBD here.
Note that \code{snpName} and \code{interval} are mutually exclusive and only one option should be specified.}

\item{interval}{a vector of length 3 containing the region to identify clusters over. This vector should contain the
chromosome ID, the start of the interval in bp and the end of the interval in bp; in this order respectively. Note that 
\code{snpName} and \code{interval} are mutually exclusive and only one option should be specified.}

\item{prop}{numeric value between 0 and 1. The minimum proportion of an interval (in bp) shared IBD between a pair of isolates in order for the pair 
to be directly connected in a cluster. For example, if \code{prop=1} then two isolates will only be connected if they are IBD over the
entire specified region, whereas if \code{prop=0.5} then two isolates will be connected if they are IBD over at least half of the interval.
The default is \code{prop=0}, which connects two isolates if they are IBD at all over the interval.}
}
\value{
A list with two objects:
\enumerate{
\item a list of disjoint clusters that contain the names of isoaltes that are part of the same network.
\item the binary IBD matrix for the specified region, where 1 represents IBD sharing by pairs over the 
specified interval and 0 otherwise.
}
}
\description{
\code{getClusters()} identifies clusters of isolates that share IBD over part of the entire specified region.
Each cluster represents a disjoint network where each isolate in the network is conneceted to all other isolates
in the network, either directly or through a series of connected isoaltes.
}

