% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_get_colour_palette.R
\name{getColourPaletteMajor}
\alias{getColourPaletteMajor}
\title{IsoRelate Colour Palette for groups}
\usage{
getColourPaletteMajor(number.colours)
}
\arguments{
\item{number.colours}{numeric. The number of colours to return.}
}
\value{
A character vector of length=\code{number.colours} containing a colour specturm.
}
\description{
\code{getColourPaletteMajor()} generates a spectrum colour palette with a specified number of colours.
}

