% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_data_processing.R
\name{getGenotypes}
\alias{getGenotypes}
\title{IsoRelate Pre-Analysis Data Processing}
\usage{
getGenotypes(ped.map, maf = 0.01, isolate.max.missing = 0.1,
  snp.max.missing = 0.1, reference.ped.map = NULL)
}
\arguments{
\item{ped.map}{A list with 2 objects:
\enumerate{
\item An object of class pedMatrix which contains the PLINK PED information as described in \code{\link{between_load_pedmap}}
\item An object of class mapMatrix which contains the PLINK MAP information as described in \code{\link{between_load_pedmap}}
}}

\item{maf}{A numeric value denoting the smallest minor allele frequency allowed in the analysis. The 
default value is 0.01.}

\item{isolate.max.missing}{A numeric value denoting the maximum proportion of missing data allowed for 
each isolate. The default value is 0.1.}

\item{snp.max.missing}{A numeric value denoting the maximum proportion of missing data allowed for each 
SNP. The default value is 0.1.}

\item{reference.ped.map}{An optional list containing reference data used to calculate population allele 
frequencies. The list has 2 objects in the same format as the input 
\code{ped.map} objects. The default value is \code{NULL} which calculates the population allele 
frequencies from the input data.}
}
\value{
A list of two objects: 
\enumerate{
\item A pedigree containing the isolates that remain after filtering. The pedigree is the first six columns 
of the PED file and these columns are
headed \code{fid, iid, pid, mid, moi} and \code{aff} respectively.
\item A data frame with the first five columns:
\enumerate{
\item Chromosome (type \code{"character"}, \code{"numeric"} or \code{"integer"})
\item SNP identifiers (type \code{"character"})
\item Genetic map distance (Morgans, M) (type \code{"numeric"})
\item Base-pair position (type \code{"integer"})
\item Population allele frequency (type \code{"integer"})
} 
where each row describes a single marker. These columns are headed \code{CHROMOSOME, MARKER, POSITION.M, POSITION.bp} and \code{FREQ} respectively. 
Columns 6 onwards contain the genotype data for each isolate, where a single column corresponds to a single isolate. These columns are 
labelled with merged family IDs and isolate IDs separated by a slash symbol (/). 
}
}
\description{
\code{getGenotypes()} performs pre-analysis data processing of PLINK formatted unphase haplotype data, 
including removal of SNPs and isolates with high proportions of missing data and SNPs with low minor 
allele frequencies. It also calculates population allele frequencies for each SNP from either the 
input dataset or a specified reference dataet.
}

