% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ibd_i_clusters.R
\name{getIBDiclusters}
\alias{getIBDiclusters}
\title{Interval Cluster Networks}
\usage{
getIBDiclusters(ped.genotypes, ibd.segments, interval = NULL, prop = 0,
  hi.clust = FALSE)
}
\arguments{
\item{ped.genotypes}{A list containing 2 objects. See the \code{Value} description in \code{\link{getGenotypes}} for more details on this input.}

\item{ibd.segments}{A data frame containing the IBD segments detected by isoRelate.
See the \code{Value} description in \code{\link{getIBDsegments}} for more details on this input.}

\item{interval}{A vector of length 3 containing the region to identify clusters over. This vector should contain the
chromosome ID, the start of the interval in base-pairs and the end of the interval in base-pairs; in this order respectively.}

\item{prop}{Numeric value between 0 and 1 (inclusive).
The minimum proportion of an interval (in base-pairs) shared IBD between a pair of isolates in order for the pair to be included in the network.
For example, if \code{prop=1} then two isolates will be included if they are IBD over the entire interval,
whereas \code{prop=0.5} will include isolates that are IBD over at least 50\% of the interval.
The default is \code{prop=0}, which includes isolates if they have an IBD segment that overlaps the interval in any way.}

\item{hi.clust}{Logical. Whether to perform hierarchical clustering using the \code{fastgreedy.community} approach in the \code{igraph} package.}
}
\value{
A list of three objects named \code{clusters}, \code{i.network} and \code{hi.clust}:
\enumerate{
\item A list where each object contains the names of isolates that form a disjoint cluster in the network.
If hierarchical clustering has been performed then the clusters may not be disjoint.
\item An \code{igraph} network used in the construction of network plots. See \url{ http://igraph.org/r/} doe more details.
\item Logical. Whether or not hierarchical clustering has been performed.
}
}
\description{
\code{getIBDiclusters()} produces a network of clusters of isolates that have been inferred IBD over a specified interval.
Isolates that are not IBD over the interval are not included in the network or output. The networks are created
using the R package \code{igraph}.
}

