% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ibd_iR.R
\name{getIBDiR}
\alias{getIBDiR}
\title{Selection Significance Statistic}
\usage{
getIBDiR(ped.genotypes, ibd.matrix, groups = NULL)
}
\arguments{
\item{ped.genotypes}{A list containing 2 objects. See the \code{Value} description in \code{\link{getGenotypes}} for more details on this input.}

\item{ibd.matrix}{A data frame containing the binary IBD information for each SNP and each pair.
See the returned \code{Value} in \code{\link{getIBDmatrix}} for more details.}

\item{groups}{A data frame with 3 columns of information:
\enumerate{
\item Family ID
\item Isolate ID
\item Group ID
}
where IBD proportions are calculated for
\enumerate{
\item all pairs of isolates within the same group
\item all pairwise-group comparisons where isolates belong to different groups
}
Group ID, for example, can be the geographic regions where the isolates were collected.
The default is \code{groups=NULL} and IBD proportions will be calculated over all pairs.}
}
\value{
A data frame the following 7 columns:
\enumerate{
\item Chromosome (type \code{"character"}, \code{"numeric"} or \code{"integer"})
\item SNP identifiers (type \code{"character"})
\item Genetic map distance (centi morgans, cM) (type \code{"numeric"})
\item Base-pair position (type \code{"integer"})
\item Population (type \code{"character"} or \code{"numeric"})
\item Subpopulation (type \code{"character"} or \code{"numeric"})
\item iR statistic (type \code{"numeric"})
\item -log10 p vlaue (type \code{"numeric"})
}
where each row describes a unique SNP.
The column 'Population' is filled with ones by default, while 'Subpopulation' contains the group ID’s from  \code{groups},
where the proportion of pairs IBD has been calculated for all pairs of isolates belonging to the same group as well as all pairs of
isolates where each isolate belongs to a different group.
If \code{groups=NULL} then 'Subpopulation' will be filled with zeros also.
The population columns have been included for plotting purposes.
The data frame is headed \code{chr, snp_id, pos_M, pos_bp, pop, subpop, iR} and \code{log10_pvalue} respectively.
}
\description{
\code{getIBDiR()} calculates a summary statistic for each SNP that can be used to assess the significance of excess IBD sharing at genomic loci,
thus identifying regions under positive selection.
First relatedness between isolates and SNP allele frequencies are accounted for, then normalization procedures are applied where we assume
our transformed summary statistic follows a chi-squared distribution with 1 degree of freedom.
This allows the calculation of –log10 (P-values) which we denote as the iR statistic.
SNPs with iR values greater than some threshold (i.e. –log10 (P-values) > –log10 (0.05)) provide evidence of positive selection.
\code{getIBDiR} can return \code{NA} iR statistics for a number of reasons, including trying to generate iR statistics when there are no IBD pairs
or when all pairs are IBD, or when only several isolates are analyzed.
}
