% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ibd_matrix.R
\name{getIBDmatrix}
\alias{getIBDmatrix}
\title{Binary IBD Matrix}
\usage{
getIBDmatrix(ped.genotypes, ibd.segments)
}
\arguments{
\item{ped.genotypes}{A list containing 2 objects. See the \code{Value} description in \code{\link{getGenotypes}} for more details on this input.}

\item{ibd.segments}{A data frame containing the IBD segments detected by isoRelate.
See the \code{Value} description in \code{\link{getIBDsegments}} for more details on this input.}
}
\value{
A data frame with the first four columns:
\enumerate{
\item Chromosome (type \code{"character"}, \code{"numeric"} or \code{"integer"})
\item SNP identifiers (type \code{"character"})
\item Genetic map distance (centi morgans, cM) (type \code{"numeric"})
\item Base-pair position (type \code{"integer"})
}
where each row describes a unique SNP. Columns 1-4 are headed \code{chr, snp_id, pos_M} and \code{pos_bp} respectively.
Columns 5 onwards contain the binary IBD information for each isolate pair, where a single column corresponds to a single pair.
These columns are labeled with merged family IDs and isolate IDs separated by a slash symbol (/). For example fid1/iid1/fid2/iid2.
}
\description{
\code{getIBDmatrix()} produces a binary matrix of IBD (1) and non-IBD (0) results for each SNP and isolate pair combination.
Each row identifies a unique SNP while each column identifies a unique isolate pair.
}

