% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_parameter_estimation.R
\name{getIBDparameters}
\alias{getIBDparameters}
\title{Between Isolates Parameter Estimation}
\usage{
getIBDparameters(ped.genotypes, number.cores = 1)
}
\arguments{
\item{ped.genotypes}{A list containing 2 objects. See the \code{Value} description in \code{\link{between_data_processing}} for more details on this input.
Note the family IDs and isolate IDs in obeject 1 of this list must match the family IDs and isolate IDs in the header of object 2 of this list.}

\item{number.cores}{Integer. The number of cores used for parallele execution.}
}
\value{
A data frame with the following eight columns:
\enumerate{
\item Family 1 ID
\item Isolate 1 ID
\item Family 2 ID
\item Isolate 2 ID
\item The number of meiosis
\item Probability of sharing 0 alleles IBD
\item Probability of sharing 1 allele IBD
\item Probability of sharing 2 alleles IBD
}
where each row describes a unique pair of isolates. The data frame is headed \code{fid.1, iid1, fid.2, iid2, m, ibd0, ibd1} and \code{ibd2} respectively.
}
\description{
\code{getIBDparameters()} estimates the number of meiosis and the probabilities of sharing 0, 1 and 2 alleles IBD between all pairwise
combinations of isolates.
}

