% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ibd_p_clusters.R
\name{getIBDpclusters}
\alias{getIBDpclusters}
\title{Genome Cluster Networks}
\usage{
getIBDpclusters(ped.genotypes, ibd.segments, prop = 1, hi.clust = FALSE)
}
\arguments{
\item{ped.genotypes}{A list containing 2 objects. See the \code{Value} description in \code{\link{getGenotypes}} for more details on this input.}

\item{ibd.segments}{A data frame containing the IBD segments detected by isoRelate.
See the \code{Value} description in \code{\link{getIBDsegments}} for more details on this input.}

\item{prop}{Numeric value between (0,1].
The minimum proportion of genome shared IBD between a pair of isolates in order for the pair to be included in the network.
For example, if \code{prop=1} then two isolates will be included if they are IBD over the entire genome, i.e. identical isolates,
whereas \code{prop=0.5} will include isolates that share at least 50\% of their genome IBD.
The default is \code{prop=1}.}

\item{hi.clust}{Logical. Whether to perform hierarchical clustering using the \code{fastgreedy.community} approach in the \code{igraph} package.}
}
\value{
A list of three objects named \code{clusters}, \code{i.network} and \code{hi.clust}:
\enumerate{
\item A list where each object contains the names of isolates that form a disjoint cluster in the network.
If hierarchical clustering has been performed then the clusters may not be disjoint.
\item An \code{igraph} network used in the construction of network plots. See \url{ http://igraph.org/r/} doe more details.
\item Logical. Whether or not hierarchical clustering has been performed.
}
}
\description{
\code{getIBDpclusters()} produces a network of clusters of isolates that share a minimum proportion of genome IBD.
Isolates that do not share a minimum proportion IBD are not included in the network or output.
The networks are created using R package \code{igraph}.
}

