% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ibd_posterior.R
\name{getIBDposterior}
\alias{getIBDposterior}
\title{IBD Posterior Probabilities}
\usage{
getIBDposterior(ped.genotypes, parameters, number.cores = 1, error = 0.001)
}
\arguments{
\item{ped.genotypes}{A list containing 2 objects.
See the \code{Value} description in \code{\link{getGenotypes}} for more details on this input.}

\item{parameters}{A data frame containing meioses and IBD probability estimates for all pairwise combinations of isolates.
See the \code{Value} description in \code{\link{getIBDparameters}} for more details on this input.}

\item{number.cores}{Positive integer. The number of cores used for parallel execution.}

\item{error}{The genotyping error rate. The default value is 0.001.}
}
\value{
A data frame with the first four columns:
 \enumerate{
\item Chromosome
\item SNP identifiers
\item Genetic map distance
\item Base-pair position
}
where each row describes a single SNP. These columns are headed \code{chr, snp_id, pos_M} and \code{pos_bp} respectively.
Columns 5 onwards contain the posterior probabilities for each pair of isolates, where a single column corresponds to one pair of isolates.
These columns are labeled with merged family IDs and isolate IDs separated by a slash symbol (/).
}
\description{
\code{getIBDposterior()} calculates the posterior probabilities of IBD sharing between pairs of isolates.
}
\examples{
\dontrun{
# calculate the posterior probability of IBD sharing
# note: this can take a while to run if there are many pairs
my_posterior <- getIBDposterior(ped.genotypes = png_genotypes,
                                parameters = png_parameters,
                                number.cores = 1,
                                error = 0.001)

head(my_posterior[,1:10])
}
}
