% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ibd_proportions.R
\name{getIBDproportion}
\alias{getIBDproportion}
\title{Proportion of Pairs IBD}
\usage{
getIBDproportion(ped.genotypes, ibd.matrix, groups = NULL)
}
\arguments{
\item{ped.genotypes}{A list containing 2 objects. See the \code{Value} description in \code{\link{getGenotypes}} for more details on this input.}

\item{ibd.matrix}{A data frame containing the binary IBD information for each SNP and each pair.
See the returned \code{Value} in \code{\link{getIBDmatrix}} for more details.}

\item{groups}{A data frame with 3 columns of information:
\enumerate{
\item Family ID
\item Isolate ID
\item Group ID
}
where, if specified, IBD proportions are calculated for
\enumerate{
\item all pairs of isolates within the same group
\item all pairwise-group comparisons where isolates belong to different groups
}
Group ID, for example, can be the geographic regions where the isolates were collected.
The default is \code{groups=NULL} and IBD proportions will be calculated over all pairs.}
}
\value{
A data frame the following 7 columns:
\enumerate{
\item Chromosome (type \code{"character"}, \code{"numeric"} or \code{"integer"})
\item SNP identifiers (type \code{"character"})
\item Genetic map distance (centi morgans, cM) (type \code{"numeric"})
\item Base-pair position (type \code{"integer"})
\item Population (type \code{"character"} or \code{"numeric"})
\item Subpopulation (type \code{"character"} or \code{"numeric"})
\item Proportion of pairs IBD (type \code{"integer"})
}
where each row describes a unique SNP.
The column \code{Population} is filled with 1s by default, while \code{Subpopulation} contains the group IDs from \code{groups},
where the proportion of pairs IBD has been calculated for all isolates belonging to the same group as well as all isolates from different groups.
If \code{groups=NULL} then \code{Subpopulation} will be filled with 1s also.
The population columns have been included for plotting purposes.
The data frame is headed \code{chr, snp_id, pos_M, pos_bp, pop, subpop} and \code{prop_ibd} respectively.
}
\description{
\code{getIBDproportion()} calculates the proportion of pairs inferred IBD at each SNP.
}
