% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ibd_summary.R
\name{getIBDsummary}
\alias{getIBDsummary}
\title{IBD Segment Summary}
\usage{
getIBDsummary(ped.genotypes, ibd.segments)
}
\arguments{
\item{ped.genotypes}{A list containing 2 objects.
See the \code{Value} description in \code{\link{getGenotypes}} for more details on this input.}

\item{ibd.segments}{A data frame containing the IBD segments detected by isoRelate.
See the \code{Value} description in \code{link{getIBDsegments}} for more details on this input.}
}
\description{
\code{getIBDsummary()} prints a brief summary of the detected IBD segments to the console.
}
