% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_locus_matrix.R
\name{getLocusMatrix}
\alias{getLocusMatrix}
\title{Between Isolates Binary IBD Matrix}
\usage{
getLocusMatrix(ped.genotypes, ibd.segments)
}
\arguments{
\item{ped.genotypes}{A list containing 2 objects. See the \code{Value} description in \code{\link{between_data_processing}} for more details on this input.
Note the family IDs and isolate IDs in obeject 1 of this list must match the family IDs and isolate IDs in the header of object 2 of this list.}
}
\value{
A data frame the first four columns:
\enumerate{
\item Chromosome (type \code{"character"}, \code{"numeric"} or \code{"integer"})
\item SNP identifiers (type \code{"character"})
\item Genetic map distance (centi morgans, cM) (type \code{"numeric"})
\item Base-pair position (type \code{"integer"})
}
where each row describes a unique SNP. These columns are headed \code{CHROMOSOME, MARKER, POSITION.cM} and \code{POSITION.bp} respectively.
Columns 5 onwards contain the binary IBD information for each isolate pair, where a single column corresponds to a single pair. 
These columns are labelled with merged family IDs and isolate IDs separated by a slash symbol (/). For example famA/ind1/famA/ind2.
}
\description{
\code{getLocusMatrix()} produces a binary matrix of IBD (1) and non-IBD (0) for each SNP and pair combination.
The number of rows is equal to the number of SNPs and the number of columns is equal to the number of pairs.
}

