% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_locus_sum.R
\name{getLocusProportion}
\alias{getLocusProportion}
\title{Between Isolates Proportion of Pairs IBD}
\usage{
getLocusProportion(ped.genotypes, locus.matrix, groups = NULL)
}
\arguments{
\item{ped.genotypes}{a list containing 2 objects. See the \code{Value} description in \code{\link{between_data_processing}} for more details on this input.
Note the family IDs and isolate IDs in obeject 1 of this list must match the family IDs and isolate IDs in the header of object 2 of this list.}

\item{groups}{a data frame with a least 3 coloumns of information:
\enumerate{
\item Family ID
\item Isolate ID
\item Group ID 1
}
Group ID, for example, can be geographic regions where the isolates were collected. A fourth column can also be included with a second level of group IDs.
If \code{groups} is specified, each isolate in the pedigree should belong to a group. If more than two levels of groups are specified, only the first
two levels will be used, i.e. columns 1 to 4 in \code{groups} will be used. The default is \code{groups=NULL} for no groups.}

\item{locus_matrix}{A data frame containing the binary IBD information for each SNP and each pair. 
See the returned \code{value} in \code{\link{between_locus_matrix}} for more details.}
}
\value{
A data frame the following 5 columns:
\enumerate{
\item Chromosome (type \code{"character"}, \code{"numeric"} or \code{"integer"})
\item SNP identifiers (type \code{"character"})
\item Genetic map distance (centi morgans, cM) (type \code{"numeric"})
\item Base-pair position (type \code{"integer"})
\item Number of IBD segments (type \code{"integer"})
}
where each row describes a unique SNP. The data frame is headed 
\code{CHROMOSOME, MARKER, POSITION.cM, POSITION.bp} and \code{prop} respectively.
}
\description{
\code{getLocusProportion()} calculates the number of IBD segments inferred at each SNP over all pairs.
}

