% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_get_overlap.R
\name{getOverlap}
\alias{getOverlap}
\title{function to find IBD that overlap interval}
\usage{
getOverlap(region.1, region.2)
}
\arguments{
\item{region.1}{IBD segment boundary}

\item{region.2}{interest interval}
}
\description{
function to find IBD that overlap interval
}
