% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{groupPairs}
\alias{groupPairs}
\title{Group Combinations for Analysis}
\usage{
groupPairs(group)
}
\arguments{
\item{group}{A character vector of all family IDs}
}
\description{
Creates a data frame containing family IDs and isolate IDs for each pair to be analysed. Each row
corresponds to a unique pair.
}

