% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{haplotypeToGenotype}
\alias{haplotypeToGenotype}
\title{Call Genotypes from Haplotype Data}
\usage{
haplotypeToGenotype(haplotypes, moi)
}
\arguments{
\item{haplotypes}{An integer matrix of haplotype data in PLINK format. A allele is denoted 1, B allele is denoted 2 and missing data is denoted 0}

\item{moi}{An integer vector of multiplicity of infection (MOI) estimates for each isoalte. Isolate MOI estimates should be ordered such that value \code{n} of
\code{moi} corresponds to column \code{n} of \code{haplotypes}.}
}
\value{
A matrix with genotype calls where columns correspond to isolates and rows correspond to SNPs
}
\description{
\code{haplotypeToGenotype} transforms PLINK haplotype data into genotype data of the form -1, 0, 1 and 2 representing missing genotypes,
homozygous reference, heterozygous and homozygous alternative respectively. Haploid isolates are coded as diploid although will not have
heterozygous genotypes.
}

