% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{IBDLabel}
\alias{IBDLabel}
\title{Internal Function}
\usage{
IBDLabel(snp_id, number_snps)
}
\arguments{
\item{snp_id}{A numeric vector of SNP identifiers for IBD segments on the chromosome of interest.}

\item{number_snps}{integer. The number of IBD SNPs for the chromosome of interest.}
}
\description{
\code{IBDLabel} is a function used to label unique IBD segments in a pair of isolates for a particular chromosome by determining 
breakpoints in IBD vs non-IBD regions. IBD segments are labelled in sequential order genome wide.
}

