% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{IBDMatrix}
\alias{IBDMatrix}
\title{Binary Matrix of IBD}
\usage{
IBDMatrix(chromosomes, positions_bp, number_pairs, ibd_pairs_colnumbers,
  ibd_chromosomes, ibd_start_bp, ibd_stop_bp)
}
\arguments{
\item{chromosomes}{A character vector containing the corresponding chromosome for each SNP}

\item{positions_bp}{A numeric vector containing the corresponding bp position for each SNP}

\item{number_pairs}{Numeric. The total number of pairs analysed}

\item{ibd_pairs_colnumbers}{A numeric vector corresponding to column numbers in the output matrix 
where each unique number refers to a unique pair with IBD inferred}

\item{ibd_chromosomes}{A character vector containing the chromosome for each detected IBD segment}

\item{ibd_start_bp}{A numeric vector containing the base-pair position for the start of each detected IBD segment}

\item{ibd_stop_bp}{A numeric vector containing the base-pair position for the end of each detected IBD segment}
}
\description{
Creates a binary matrix of IBD (1) and non-IBD (0) with each row representing a single SNP
and each column representing a unique pair. The number of rows is equal to the total 
number of SNPs and the number of columns is equal to the number of pairs.
}

