% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_meiosis.R
\name{IBDparameters}
\alias{IBDparameters}
\title{IBDparameters() calculates IBD probabilities then estimates the number of meiosis for an isolate pair.}
\usage{
IBDparameters(genotypes, pop_allele_freqs, gender_1, gender_2)
}
\arguments{
\item{genotypes}{An integer matrix of genotype calls for a pair of isolates. Each coloumn represents and isolate and each row represents a SNP.}

\item{pop_allele_freqs}{A numeric vector of population allele frequencies for each SNP.}

\item{gender_1}{An integer denoting the MOI value of isoalte 1.}

\item{gender_2}{An integer denoting the MOI value of isoalte 2.}
}
\value{
A vector of 4 values representing the number of meiosis and the probabilities of sharing 0, 1 and 2 alleles IBD respectively.
}
\description{
IBDparameters() calculates IBD probabilities then estimates the number of meiosis for an isolate pair.
}

