% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_segment_summary.R
\name{IBDTable}
\alias{IBDTable}
\title{Internal Function}
\usage{
IBDTable(ibd.results)
}
\arguments{
\item{ibd.results}{A data frame containing family ID and isolate ID for isolate 1, family ID and isolate ID for isolate 2, 
numeric SNP identifiers, chromosome identifiers, genetic map positions of SNPs in Morgans (M) and base-pairs (bp) and the Viterbi results respectively, 
for all SNPs.}
}
\value{
A data frame containing a summary of all IBD segments inferred for this pair of isolates. The data frame contains
the following columns:
\enumerate{
\item Family 1 ID
\item Isolate 1 ID
\item Family 2 ID
\item Isolate 2 ID
\item Chromosome
\item SNP identifier
\item Start SNP
\item End SNP
\item Start position bp
\item End position bp
\item Start position M
\item End position M
\item Number of SNPs
\item Length bp
\item Length M
\item IBD status (1 = 1 allele shared IBD, 2 = 2 alleles shared IBD)
}
}
\description{
\code{IBDTable()} produces summaries of detected IBD segments for a single pair of isolates. These summaries include the genetic 
map start and end
of IBD segments in bp, cM and SNP identifiers; and lengths of IBD segments in bp, cM and SNPs.
}
