% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_iR_function.R
\name{iRfunction}
\alias{iRfunction}
\title{Internal Function}
\usage{
iRfunction(locus.matrix, frequency, plot_qq = FALSE)
}
\arguments{
\item{locus.matrix}{A data frame containing binary IBD values. See \code{\link{getLocusMatrix}} for more details.}

\item{frequency}{A vector of population allele frequencies for each SNP.}
}
\description{
\code{iRfunction()} calculates the iR statistic used to assess the significance of excess IBD sharing at loci
in the genome. The final statistic, -log10 (P-values), is returned for each SNP.
}

