% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isolatePairs}
\alias{isolatePairs}
\title{Pair Combinations for Analysis}
\usage{
isolatePairs(fid, iid)
}
\arguments{
\item{fid}{A character vector of all family IDs}

\item{iid}{A character vector of all individual ID}
}
\description{
Creates a data frame containing family IDs and isolate IDs for each pair to be analysed. Each row
corresponds to a unique pair.
}

