% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_load_pedmap.R
\name{loadPEDMAP}
\alias{loadPEDMAP}
\title{IsoRelate Load in PLINK PED and MAP files}
\usage{
loadPEDMAP(ped, map)
}
\arguments{
\item{ped}{A character string containing the path and filename of the input PLINK PED file. The PED file is 
a white-space (space or tab) delimited file
with the first six columns:
\enumerate{
\item Family ID
\item Isolate ID
\item Paternal ID
\item Maternal ID
\item Multiplicity of infection (MOI) (1 = single infection or haploid, 2 = multiple infections or diploid)
\item Phenotype (1 = unaffected, 2 = affected, 0 = unknown)
}
The IDs are alphanumeric: the combination of family and individual ID should uniquely identify an isolate. The phenotype coulmn is not used
in IsoRelate analyses however is required for completeness of a standard pedigree. Genotypes (column 7 onwards) should
also be white-spaced delimited with the A and B alleles coded as 1 and 2 respectively and missing data coded as 0. All SNPs (whether haploid or not)
must have two alleles specified. For haploid chromosomes, genotypes should be specified as homozygous. Either both alleles should be missing (i.e. 0)
or neither. No header row should be given.}

\item{map}{A character string containing the path and filename of the input PLINK MAP file. A MAP file contains exactly four coloums of information:
\enumerate{
\item Chromosome
\item SNP identifier
\item Genetic map distance (centi morgans, cM)
\item Base-pair position
}
where each row describes a single marker. Genetic map distance and base-pair positions are expected to be positive values. The MAP file must
be ordered by increasing chromosomes and positions. SNP identifiers can contain any characters expect spaces or tabs; also you should avoid
* symbols in the names. The MAP file must contain as many markers as are in the PED file. No header row should be given.}
}
\value{
A list of two objects: 
\enumerate{
\item An object of class \code{pedMatrix} containing the data from \code{ped}.
\item An object of class \code{mapMatrix} containing the data from \code{map}. 
}
}
\description{
\code{loadPEDMAP()} reads in a PLINK PED and PLINK MAP file using \code{fread} from the \code{data.table} package.
}

