% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_merge_lists.R
\name{merge_lists}
\alias{merge_lists}
\title{Internal Function}
\usage{
merge_lists(A, B)
}
\arguments{
\item{A}{List with 2 objects for one pair. Object 1 is the IBD summaries for a pair and object 2 is the IBD posterior probabilities.}

\item{B}{List with 2 objects for another pair. Object 1 is the IBD summaries for a pair and object 2 is the IBD posterior probabilities.}
}
\value{
A list with 2 objects containing merged lists from \code{A} and \code{B}    above.
}
\description{
\code{merge_lists()} is a function used to merge summary IBD results for multiple pairs when running the IBD analysis on 
multiple cores
}

