% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_plot_ibd_zoom.R
\name{overlappingSegment}
\alias{overlappingSegment}
\title{Internal Function}
\usage{
overlappingSegment(a, b)
}
\arguments{
\item{a}{A vector of containing the start and end positions of an interval in bp}

\item{b}{A vector of containing the start and end positions of an interval in bp}
}
\description{
\code{overlappingSegment()} returns the maximum start of two regions and the minimum end of two regions.
}

