% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_plot_clusters.R
\name{plotClusters}
\alias{plotClusters}
\title{Between Isolates Clusters}
\usage{
plotClusters(ped.genotypes, clusters, groups = NULL, vertex.size = 4,
  vertex.name = FALSE, vertex.frame.color = "white", add.legend = TRUE,
  ...)
}
\arguments{
\item{ped.genotypes}{a list containing 2 objects. See the \code{Value} description in \code{\link{between_data_processing}} for more details on this input.
Note the family IDs and isolate IDs in obeject 1 of this list must match the family IDs and isolate IDs in the header of object 2 of this list.}

\item{clusters}{a data frame containing the binary IBD information for each SNP and each pair. 
See the returned \code{value} in \code{\link{between_locus_matrix}} for more details.}

\item{groups}{a character string of length 1 giving the name of a SNP to find clusters of isolates who are IBD here.
Note that \code{snpName} and \code{interval} are mutually exclusive and only one option should be specified.}

\item{vertex.size}{numeric value between 0 and 1. The minimum proportion of an interval (in bp) shared IBD between a pair of isolates in order for the pair 
to be directly connected in a cluster. For example, if \code{prop=1} then two isolates will only be connected if they are IBD over the
entire specified region, whereas if \code{prop=0.5} then two isolates will be connected if they are IBD over at least half of the interval.
The default is \code{prop=0}, which connects two isolates if they are IBD at all over the interval.}

\item{vertex.name}{logical. Whether to add isolate names to the nodes. Default is \code{vertex.name=FALSE}.}

\item{vertex.frame.color}{character string. A single colour that outlines vertices. Default is \code{vertex.fame.color="white"}.}

\item{add.legend}{logical. Whether to plot a legend. This only applies when isolates have MOI > 1 or \code{groups} is specified.}

\item{...}{other plotting parameters allowed by \code{igraph}.}

\item{groups}{a vector of length 3 containing the region to identify clusters over. This vector should contain the
chromosome ID, the start of the interval in bp and the end of the interval in bp; in this order respectively. Note that 
\code{snpName} and \code{interval} are mutually exclusive and only one option should be specified.}
}
\value{
A list of disjoint clusters that contain the names of isoaltes that are part of the same network.
The last object in the list is the binary IBD matrix for the specified region, where 1 represents IBD sharing by
pairs over the specified interval and 0 otherwise.
}
\description{
\code{getClusters()} identifies clusters of isolates that share IBD over part of the entire specified region.
Each cluster represents a disjoint network where each isolate in the network is conneceted to all other isolates
in the network, either directly or through a series of connected isoaltes.
}

