% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_plot_ibd.R
\name{plotIBD}
\alias{plotIBD}
\title{Internal Function}
\usage{
plotIBD(ibd.segments, ped.genotypes, chrom = NULL,
  ibd.colours = c("dodgerblue", "darkgoldenrod1"), pairs.per.page = 50,
  iid.only = FALSE, fid.only = FALSE)
}
\arguments{
\item{ibd.segments}{A data frame with inferred IBD tracts. See the \code{Value} description in \code{\link{between_ibd_segments}} 
for more details}

\item{ped.genotypes}{A list containing 2 objects. See the \code{Value} description in \code{\link{between_processing}} for more details on this input.}

\item{chrom}{A vector of chromosomes to plot. Can be either numeric or character.}

\item{ibd.colours}{A vector with two values. These can be either numeric or the name of colours to be used in the IBD plot (must be vaild colour
names).}

\item{pairs.per.page}{Integer. The maximum number of pairwise results to plot on each page of the PDF.}

\item{iid.only}{Use only the isolate IDs on the figure legend of the plot. If using, the isolate IDs must be unique for each isolate.
\code{iid.only} and \code{fid.only} are mutually exclusive.}

\item{fid.only}{Use only the family IDs on the figure legend of the plot. If using, the family IDs must be unique for each isolate.}
}
\description{
\code{plotIBD()} plots all IBD segments for all chromosomes in one figure. A maximum of 50 pairs are plotted in one figure and remaining
pairs are plotted on subsequent pdf pages with a maximum of 50 pairs in one figure.
}

