% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_plot_average.R
\name{plotIBDaverage}
\alias{plotIBDaverage}
\title{Between Isolates Average Sharing Plot}
\usage{
plotIBDaverage(ped.genotypes, ibd.segments, groups = NULL, number.cores = 1)
}
\arguments{
\item{locus.matrix}{A list containing 2 objects. See the \code{Value} description in \code{\link{between_data_processing}} for more details on this input.
Note the family IDs and isolate IDs in obeject 1 of this list must match the family IDs and isolate IDs in the header of object 2 of this list.}
}
\description{
\code{plotIBDaverage()} plots the proportion of pairs IBD for each SNP
}

