% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ibd_clusters.R
\name{plotIBDclusters}
\alias{plotIBDclusters}
\title{Plot Cluster Networks}
\usage{
plotIBDclusters(ped.genotypes, clusters, groups = NULL, vertex.color = NULL,
  vertex.frame.color = "white", vertex.size = 4, vertex.name = FALSE,
  edge.color = "gray60", edge.width = 0.8, mark.border = "white",
  mark.col = "gray94", add.legend = TRUE, legend.x = -1.5,
  legend.y = -0.25, layout = NULL, return.layout = FALSE)
}
\arguments{
\item{ped.genotypes}{A list containing 2 objects. See the \code{Value} description in \code{\link{getGenotypes}} for more details on this input.}

\item{clusters}{A named list of three objects containing network information.
See the \code{Value} description in either \code{\link{getIBDiclusters}} or \code{\link{getIBDpclusters}} for more details on this input.}

\item{groups}{A data frame with 3 columns of information:
\enumerate{
\item Family ID
\item Isolate ID
\item Group ID
}
Group ID, for example, can be the geographic regions where the isolates were collected.
If \code{groups} is specified then each isolate in the pedigree must belong to a group.
Vertices in the network will be colored according to group allocation.
The default is \code{groups=NULL} and all vertices will have the same color.}

\item{vertex.color}{A vector of characters or numeric values of the vertex colors in the network.
If \code{groups} is specified then \code{vertex.color} should contain the same number of colors as unique groups.}

\item{vertex.frame.color}{Character string or numeric value. A single color that will be used as the vertex border. Default is \code{vertex.fame.color="white"}.}

\item{vertex.size}{Numeric value indicating the size of the vertices in the network. Default is \code{vertex.size=4}.}

\item{vertex.name}{Logical. Whether to add isolate names to the vertices. Default is \code{vertex.name=FALSE}.}

\item{edge.color}{Character string or numeric value. A single color to be used for all edges. Default is \code{edge.color="gray60"}.}

\item{edge.width}{Numeric. A single value indicating the width of the edges. Default is \code{edge.width=0.8}.}

\item{mark.border}{Character string or numeric value. A single color to be used for all borders in hierarchical clustering groups. Default is \code{mark.border="white"}.}

\item{mark.col}{Character string or numeric value. A single color to be used to fill hierarchical clustering groupings. Default is \code{mark.col="gray94"}.}

\item{add.legend}{Logical. Whether to include a legend in the plot. Default is \code{add.legend=TRUE}.}

\item{legend.x}{Numerical. A single value indicating the x-coordinate of the legend, with default \code{legend.x=-1.5}.}

\item{legend.y}{Numerical. A single value indicating the y-coordinate of the legend, with default \code{legend.y=-0.25}.}

\item{layout}{A matrix containing the x and y coordinates of the vertices, generated using the Fruchterman-Reingold force-directed layout.}

\item{return.layout}{Logical. Whether or not to return the layout matrix (vertex positions) in the network.
This layout can be used as the input for the parameter \code{layout} to avoid different network configurations each time \code{plotClusters()}
is run on the same network.}
}
\description{
\code{plotIBDclusters()} Produces a figure of an isoRelate cluster network, where unique isolates are represented by vertices and a line is
drawn between two vertices if the isolates have been inferred IBD via the criteria specified in either \code{getIBDiclusters} or
\code{getIBDpclusters}.The networks are created using the R package \code{igraph}.
}
