% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_locus_plot.R
\name{plotIBDglobal}
\alias{plotIBDglobal}
\title{Between Isolates Loci Sum Plot}
\usage{
plotIBDglobal(locus.proportions, interval = NULL, genes = NULL)
}
\arguments{
\item{locus.proportions}{a data frame containing the binary IBD information for each SNP and each pair. 
See the returned \code{value} in \code{\link{between_locus_matrix}} for more details.}

\item{interval}{a vector of length 3 containing the region to identify clusters over. This vector should contain the
chromosome ID, the start of the interval in bp and the end of the interval in bp; in this order respectively. Note that 
\code{snpName} and \code{interval} are mutually exclusive and only one option should be specified.}

\item{genes}{a data frame with at least 5 columns of information:
\enumerate{
\item chr
\item name
\item start
\item end
\item strand
}
This data frame does not have to be in a specific order, however it must contain all of the above information with identical labels.}
}
\description{
\code{plotIBDglobal()} plots the proportion of pairs IBD for each SNP
}

