% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ibd_iR.R
\name{plotIBDiR}
\alias{plotIBDiR}
\title{Plot iR Statistics}
\usage{
plotIBDiR(ibd.iR, interval = NULL, annotation.genes = NULL,
  annotation.genes.color = NULL, highlight.genes = NULL,
  highlight.genes.labels = TRUE, highlight.genes.color = NULL,
  highlight.genes.alpha = 0.1, point.size = 1, point.color = NULL,
  add.rug = FALSE, plot.title = NULL, add.legend = FALSE,
  facet.label = TRUE, facet.scales = "fixed")
}
\arguments{
\item{ibd.iR}{A data frame containing the iR summary statistics for each SNP.
See the returned \code{Value} in \code{\link{getIBDiR}} for more details.
If multiple subpopulations are specified (column name "subpop") then the iR statistics for each subpopulation will be plotted
on a separate facet in the figure (see \url{http://docs.ggplot2.org/current/facet_grid.html} on faceting).
If there are many subpopulations (>8) it may be better to plot subsets of the subpopulations as apposed to all subpopulations in a single figure.
If multiple populations (column name "pop") are specified then only the first population will be included in the figure.
Genomic locations of annotation genes can be included in the figure and specific intervals can be highlighted.}

\item{interval}{A vector of length 3 containing the genomic locations of a specific region to plot. This vector should contain the
chromosome ID, the start of the interval in base-pairs and the end of the interval in base-pairs; in this order respectively.
The default is \code{interval=NULL} which will plot iR statistics over all chromosomes in \code{ibd.iR}.}

\item{annotation.genes}{A data frame containing information on annotation genes to be included in the figure.
This data frame must have at least 5 columns of information:
\enumerate{
\item Chromosome (type \code{"numeric"} or \code{"integer"})
\item Gene name (type \code{"character"})
\item Start location of the gene in base-pairs (type \code{"numeric"} or \code{"integer"})
\item End location of the gene in base-pairs (type \code{"numeric"} or \code{"integer"})
\item Gene strand (+ or -) (type \code{"character"})
}
\code{annotation.genes} must contain the following headers \code{chr, name, start, end} and \code{strand}.
This data frame does not have to be in a specific order, however it must contain all of the above information
with respective labels. The default is {annotation.genes=NULL}.}

\item{annotation.genes.color}{A vector of characters or numeric values containing the two colors according to gene stand (positive (+) or negative (-))}

\item{highlight.genes}{A data frame containing information of genes or regions to highlight.
The data frame must have at least 4 columns of information:
\enumerate{
\item Chromosome (type \code{"numeric"} or \code{"integer"})
\item Gene name (type \code{"character"})
\item Start location of the gene in base-pairs (type \code{"numeric"} or \code{"integer"})
\item End location of the gene in base-pairs (type \code{"numeric"} or \code{"integer"})
}
\code{highlight.genes} should contain the following headers \code{chr, name, start} and \code{end}.
This data frame does not have to be in a specific order, however it must contain all of the above information
with respective labels. The default is {highlight.genes=NULL}.}

\item{highlight.genes.labels}{Logical. Whether to include gene names as labels in the figure. The default is \code{highlight.genes.labels=FALSE}.}

\item{highlight.genes.color}{Character string or numeric value. A single color that will be used to highlight a region/gene. The default is \code{highlight.genes.color=NULL}.}

\item{highlight.genes.alpha}{Numeric. A single value between 0 and 1 indicating the gene color transparency. The default is \code{highlight.genes.alpha=0.1}.}

\item{point.size}{Numeric. The size of the points in the figures. The default is \code{point.size=1}.}

\item{point.color}{A vector of characters or numeric values denoting the color of points to be plotted.
If there are multiple subpopulations then the number of colors specified should equal the number of subpopulations.
The default is \code{point.color=NULL} which will use isoRelate default colors.}

\item{add.rug}{Logical. Whether to include SNP positions as a rug in the figure. The default is \code{add.rug=FALSE}}

\item{plot.title}{A character string of a title to be added to the figure The default is \code{plot.title=NULL} which does not add a title to the plot.}

\item{add.legend}{Logical. Whether a legend containing subpopulation information should be plotted. The default is \code{add.legend=FALSE}.}

\item{facet.label}{Logical. Whether to include facet labels if multiple subpopulations (column name "subpop") are specified.}

\item{facet.scales}{A character string of either \code{"fixed"}, \code{"free"}, \code{"free_x"} or \code{"free_y"} specifying the facet axis-scales.
See \url{http://docs.ggplot2.org/current/facet_grid.htmlfacet.scales} for more information on this parameter. The default is \code{facet.scales="fixed"}}
}
\description{
\code{plotIBDiR()} plots the -log10 (P-values) used to assess the significance of excess IBD sharing.
}

