% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_ibd_plot.R
\name{plotIBDpairwise}
\alias{plotIBDpairwise}
\title{Between Isolates IBD Plotting}
\usage{
plotIBDpairwise(ped.genotypes, ibd.segments, chromosomes = NULL,
  zoom = NULL, ibd.colours = c("dodgerblue", "darkgoldenrod1"),
  minimum.length.bp = 0, pairs.per.page = 50, iid.only = FALSE,
  fid.only = FALSE)
}
\arguments{
\item{ped.genotypes}{A list containing 2 objects. See the \code{Value} description in \code{\link{between_processing}} for more details on this input.}

\item{ibd.segments}{A data frame with inferred IBD tracts. See the \code{Value} description in \code{\link{between_ibd_segments}} 
for more details. This data frame can be alterted prior to plotting, for example subsetting by specific combinations of isolate pairs.}

\item{chromosomes}{A vector of chromosomes to plot. Can be either numeric or character.}

\item{zoom}{A data frame containing three columns:
\enumerate{
\item Chromosome
\item Start location bp
\item End location bp
}
More than one region can be specified per chromosome however they will be plotted in spearate figures.}

\item{ibd.colours}{A vector with two values. These can be either numeric or the name of colours to be used in the IBD plot (must be vaild colour
names).}

\item{minimum.length.bp}{The minimum length of an IBD segment in bp to be included in the summary figure. The default value is 0 bp.}

\item{pairs.per.page}{The maximum number of pairwise results to plot on each page of the PDF.}

\item{iid.only}{Use only the isolate IDs on the figure legend of the plot. If using, the isolate IDs must be unique for each isolate.
\code{iid.only} and \code{fid.only} are mutually exclusive.}

\item{fid.only}{Use only the family IDs on the figure legend of the plot. If using, the family IDs must be unique for each isolate.}
}
\description{
\code{plotIBDpairwise()} produces figures of inferred IBD segments. Regions of IBD are represented by coloured rectangles, where 
\code{IBD=1} (one allele shared IBD) and \code{IBD=2} (two alleles shared IBD) regions are distingued by different colours. 
The user can choose which chromosomes they wish to plot as well as zoom in on specific regions of interest.
}

