% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ibd_segments.R
\name{plotIBDsegments}
\alias{plotIBDsegments}
\title{Plot IBD Segments}
\usage{
plotIBDsegments(ped.genotypes, ibd.segments, interval = NULL,
  annotation.genes = NULL, annotation.genes.color = NULL,
  highlight.genes = NULL, highlight.genes.labels = TRUE,
  highlight.genes.color = NULL, highlight.genes.alpha = 0.1,
  segment.height = 0.5, segment.color = NULL, number.per.page = NULL,
  fid.label = TRUE, iid.label = TRUE, ylabel.size = 9, add.rug = FALSE,
  plot.title = NULL, add.legend = TRUE)
}
\arguments{
\item{ped.genotypes}{A list containing 3 objects. See the \code{Value} description in \code{\link{getGenotypes}} for more details on this input.}

\item{ibd.segments}{A data frame containing the IBD segments inferred from pairs of isolates
See the returned \code{value} in \code{\link{getIBDsegments}} for more details.}

\item{interval}{A vector of length 3 containing the genomic locations of a specific region to plot. This vector should contain the
chromosome ID, the start of the interval in base-pairs and the end of the interval in base-pairs; in this order respectively.
The default is \code{interval=NULL} which will plot the segments over all chromosomes.}

\item{annotation.genes}{A data frame containing information on annotation genes to be included in the figure.
This data frame must have at least 5 columns of information:
\enumerate{
\item Chromosome (type \code{"numeric"} or \code{"integer"})
\item Gene name (type \code{"character"})
\item Start location of the gene in base-pairs (type \code{"numeric"} or \code{"integer"})
\item End location of the gene in base-pairs (type \code{"numeric"} or \code{"integer"})
\item Gene strand (+ or -) (type \code{"character"})
}
\code{annotation.genes} must contain the following headers \code{chr, name, start, end} and \code{strand}.
This data frame does not have to be in a specific order, however it must contain all of the above information
with respective labels. The default is {annotation.genes=NULL}.}

\item{annotation.genes.color}{A vector of characters or numeric values containing the two colors representing gene stand (positive (+) or negative (-))}

\item{highlight.genes}{A data frame containing information of genes or regions to highlight.
The data frame must have at least 4 columns of information:
\enumerate{
\item Chromosome (type \code{"numeric"} or \code{"integer"})
\item Gene name (type \code{"character"})
\item Start location of the gene in base-pairs (type \code{"numeric"} or \code{"integer"})
\item End location of the gene in base-pairs (type \code{"numeric"} or \code{"integer"})
}
\code{highlight.genes} should contain the following headers \code{chr, name, start} and \code{end}.
This data frame does not have to be in a specific order, however it must contain all of the above information
with respective labels. The default is {highlight.genes=NULL}.}

\item{highlight.genes.labels}{Logical. Whether to include gene names as labels in the figure. The default is \code{highlight.genes.labels=FALSE}.}

\item{highlight.genes.color}{Character string or numeric value. A single color that will be used to highlight a region/gene. The default is \code{highlight.genes.color=NULL}.}

\item{highlight.genes.alpha}{Numeric. A single value between 0 and 1 indicating the gene color transparency. The default is \code{highlight.genes.alpha=0.1}.}

\item{segment.height}{A numeric value giving the hight of IBD segment blocks, such that 0 < segment.height <= 1. The default is \code{segment.height=0.5}.}

\item{segment.color}{A vector of characters or numeric values denoting the color of the segments to be plotted.
Two colors must be specified, one for segments with 1 allele IBD and one for segments with 2 alleles IBD.}

\item{number.per.page}{A numeric value indicating the maximum number of IBD pairs to plot in a single graphics window. The default is
\code{number.per.page=NULL} which will plot all IBD pairs in a single window. This may not be ideal when there are many IBD pairs. If
\code{number.per.page} is set, it is recommended to plot the output to a file as opposed to the R plotting window.}

\item{fid.label}{Logical. If \code{fid.label=TRUE}, family IDs will be included in the y-axis labels; otherwise family IDs will be omitted.
The default is \code{add.fid.name=TRUE}.}

\item{iid.label}{Logical. If \code{iid.label=TRUE}, isolate IDs will be included in the y-axis labels; otherwise isolate IDs will be omitted.
The default is \code{add.iid.name=TRUE}.}

\item{ylabel.size}{A numeric value indicating the size of the y-axis labels if drawn. The default is \code{ylabel.size=9}.}

\item{add.rug}{Logical. Whether to include SNP positions as a rug in the figure. The default is \code{add.rug=FALSE}}

\item{plot.title}{A character string of a title to be added to the figure The default is \code{plot.title=NULL} which does not add a title to the plot.}

\item{add.legend}{Logical. If \code{add.legend=TRUE}, a legend specifying the IBD status (1 allele IBD or 2 alleles IBD) will be included. The default is \code{add.legend=TRUE}.}
}
\description{
\code{plotIBDsegments()} plots IBD segments for pairs across the genome. IBD segments are depicted by colored blocks.
}
