% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png_parameters.R
\docType{data}
\name{png_parameters}
\alias{png_parameters}
\title{IBD Parameter Estimates For The Papua New Guinea Dataset}
\format{A data frame with
 \describe{
\item{fid1}{Family 1 ID}
\item{iid1}{Isolate 1 ID}
\item{fid2}{Family 2 ID}
\item{iid2}{Isolate 2 ID}
\item{m}{The estimated number of meiosis separating each pair of isolates}
\item{ibd0}{The estimated proportion of genome with 0 alleles IBD}
\item{ibd1}{The estimated proportion of genome with 1 allele IBD}
\item{ibd2}{The estimated proportion of genome with 2 alleles IBD}
}}
\usage{
png_parameters
}
\description{
The parameter estimates inferred using isoRelate with the
parameter settings as in the Vignette.
}
\keyword{datasets}
