% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png_pedmap.R
\docType{data}
\name{png_pedmap}
\alias{png_pedmap}
\title{Papua New Guinea Plasmodium Falciparum Dataset}
\format{List with two objects
\describe{
\item{PED}{PLINK PED data frame. See \code{\link{getGenotypes}} for details.}
\item{MAP}{PLINK MAP data frame. See \code{\link{getGenotypes}} for details.}
}
See \code{\link{getGenotypes}} for details.}
\usage{
png_pedmap
}
\description{
A PED/MAP file containing gneotype information for 38 isolates
from Madang, Papua New Guinea. This dataset was released as part
of the MalariaGEN Pf3k consortium in VCF file format, which underwent
extensive data processing to result in the final SNP set in the
PED/MAP file. This data, and more, is available from
\url{https://www.malariagen.net/projects/pf3k}.
}
\keyword{datasets}
