% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reest_transition}
\alias{reest_transition}
\title{Re-estimate the transition probabilities}
\usage{
reest_transition(xi, gamma)
}
\arguments{
\item{xi}{A numeric matrix of xi values}

\item{gamma}{A numeric matrix of gamma values (posterior probabilities)}
}
\description{
Re-estimate the transition probabilities
}

