% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roundDecimal}
\alias{roundDecimal}
\title{Round digits to specified decimal places}
\usage{
roundDecimal(number, digits)
}
\arguments{
\item{number}{A number to round}

\item{number}{The number of digits to round to}
}
\description{
Round digits to specified decimal places
}

