% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{transitionProbHD}
\alias{transitionProbHD}
\title{The transition probabilities for 1 haploid and 1 diploid chromosome}
\usage{
transitionProbHD(omega_0, meiosis, dist_cM, ibd_current, ibd_previous)
}
\arguments{
\item{omega_0}{The probability of sharing 0 alleles IBD}

\item{meiosis}{The number of meiosis separating the two isoaltes}

\item{dist_cM}{The genetic map distance (cM) between SNP i and SNP j}

\item{ibd_current}{The IBD state of SNP j}

\item{ibd_previous}{The IBD state of SNP i}
}
\description{
The transition probabilities for 1 haploid and 1 diploid chromosome
}

