% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{transMatrix}
\alias{transMatrix}
\title{The transition probabilities for 2 haploid chromosomes}
\usage{
transMatrix(dist, transParam)
}
\arguments{
\item{dist}{The distance between adjacent SNPs}

\item{transParam}{A matrix of transition rate probabilities}
}
\description{
The transition probabilities for 2 haploid chromosomes
}

