% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trueGenotypes}
\alias{trueGenotypes}
\title{Matrices of all possible genotype combinations between pairs, given MOI}
\usage{
trueGenotypes(gender_1, gender_2)
}
\arguments{
\item{gender_1}{The MOI estimate of isolate 1}

\item{gender_2}{The MOI estimate of isolate 2}
}
\description{
Matrices of all possible genotype combinations between pairs, given MOI
}

